Temperature and memory logging script
Logs temp and memory to file with timestamped name.
flags give limits for ram and temperature warnings and errors to allow
for more granular monitoring and other steps too.
will also spam console if warnings or errors are hit due to temperature.
currently works with temperature, needs a little work to make memory
work fine.

Not convinced it should live in this folder, but don't have a suggestion
of a better one.